+++ src/node.cpp
@@ -1,8 +1,8 @@
 #include "node.hpp"
 
-qt::node::node(int x, int y, int w, int h) : x(x), y(y), w(w), h(h) {}
+qt::node::node(float x, float y, float w, float h) : x(x), y(y), w(w), h(h) {}
 
-qt::node::node(point &p, int w, int h) : x(p.x), y(p.y), w(w), h(h) {}
+qt::node::node(point &p, float w, float h) : x(p.x), y(p.y), w(w), h(h) {}
 
 bool qt::node::containsPoint(point &p){
 	if ((p.x >= x - w) && (p.y >= y - h) &&
